requirements.txt
asgiref==3.7.2
Django==5.0.3
django-admin-searchable-dropdown==1.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==24.2
django-jazzmin==2.6.1
django-simple-history==3.5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
